requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.1
beanie==1.25.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
pyautogen[all]==0.2.20
groq==0.4.0
transformers[torch]==4.40.0
torch==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.1.1
httpx==0.27.0
langchain==0.1.16 # Added for potential LangGraph use
langgraph==0.0.37 # Added for potential LangGraph use

# Note: Ensure you have a compatible CUDA version if using GPU with torch/transformers